requirements.txt
streamlit
langdetect
sentence-transformers
faiss-cpu
torch
transformers